Create description
Map matching is a step to project a GPS point to an edge correctly in a road network.  It is a pre-process for many LBS applications. There are many famous methods, for example ST-matching. Though ST-matching is a good method for people to follow, it will cost many time to perform the temporal analysis which need to find the shortest route. 

In this project, I will give two code versions. In the first verson, I do not use temporal analysis, but spatial analysis, which is an aggressive version and may correspond to lower accuracy and less time. The second verson is complete one, which I will use the temporal anaysis in.